requirements.txt
fastapi
uvicorn
requests
python-dotenv
pandas
numpy
implicit
scipy